Improve exit process
