+++ src/user/pspmoduleinfo.h
@@ -146,6 +146,7 @@ enum PspModuleInfoAttr
 
 /* Disable the use of newlib, getting a minimal binary. */
 #define PSP_DISABLE_NEWLIB() \
-	void __libcglue_init(int argc, char *argv[]) {}
+	void __libcglue_init(int argc, char *argv[]) {} \
+	void __libcglue_deinit() {}
 
 #endif /* PSPMODULEINFO_H */
